// src/ui-canvas/uic_ui_canvas/types/ComponentType.enum.ts
export enum ComponentType {
  Txt = "txt",
  Cmb = "cmb",
  Btn = "btn",
  Txa = "txa",
  Rbtn = "rbtn",
  IRbtn = "irbtn",
  Cbox = "cbox",
  Date = "date",
  Time = "time",
  Lbl = "lbl",
  Icbox = "icbox",
  File = "file",
  Hlink = "hlink",
  Img = "img",
  Tbl = "tbl",
  Grp = "grp",
  Ytube = "ytube",
}